+++ src/components/contractObligations/contractObligation.tsx
@@ -94,12 +94,12 @@ const ObligationsComponent = ({
       }));
       if (!completed) {
         // show yes no alert
-        const shouldContinue = window.confirm(
-          "Are you sure you want to mark this promise as incomplete?",
-        );
-        if (!shouldContinue) {
-          return;
-        }
+        // const shouldContinue = window.confirm(
+        //   "Are you sure you want to mark this promise as incomplete?",
+        // );
+        // if (!shouldContinue) {
+        //   return;
+        // }
       } else {
         const hasPartner = partnersData && partnersData.length > 0;
         const partnerName = hasPartner
